chore: share functionality using composition
