+++ misc/logger/src/serverChunkLogger.ts
@@ -1,66 +1,34 @@
 import { MAX_LOG_SIZE_IN_BYTES_DEFAULT } from "./constants";
 import { Writable } from "stream";
-import { LogLinkedList } from "./linkedList";
 import type { LoggerOptions } from 'pino'
-import { levels } from 'pino'
+import BaseChunkLogger from "./baseChunkLogger";
 
 export default class ServerChunkLogger extends Writable {
-  private logs: LogLinkedList;
-  private MAX_LOG_SIZE_IN_BYTES: number;
-  private level: LoggerOptions['level']
-  private levelValue: number;
+  private baseChunkLogger: BaseChunkLogger
 
   public constructor(level: LoggerOptions['level'], MAX_LOG_SIZE_IN_BYTES: number = MAX_LOG_SIZE_IN_BYTES_DEFAULT) {
     super({ objectMode: true });
 
-    this.level = level ?? 'error';
-    this.levelValue = levels.values[this.level];
-
-    this.MAX_LOG_SIZE_IN_BYTES = MAX_LOG_SIZE_IN_BYTES;
-    this.logs = new LogLinkedList();
-  }
-
-  public forwardToConsole(chunk: any) {
-    if(chunk.level === levels.values['error']) {
-      console.error(chunk)
-    }
-    else if (chunk.level === levels.values['warn']) {
-      console.warn(chunk)
-    }
-    else {
-      console.log(chunk)
-    }
+    this.baseChunkLogger = new BaseChunkLogger(level, MAX_LOG_SIZE_IN_BYTES)
   }
 
   public _write(chunk: any): void {
-    this.logs.append(
-      JSON.stringify({
-        timestamp: new Date().toISOString(),
-        log: chunk,
-      }),
-    );
-
-    // Based on https://github.com/pinojs/pino/blob/master/lib/constants.js
-    if(chunk.level >= this.levelValue) {
-      this.forwardToConsole(chunk)
-    }
-
-    if (this.logs.size >= this.MAX_LOG_SIZE_IN_BYTES) {
-      this.logs.shift();
-    }
+    this.baseChunkLogger.appendToLogs(chunk)
   }
 
   public getLogs() {
-    return this.logs;
+    return this.baseChunkLogger.getLogs();
   }
 
   public clearLogs() {
-    this.logs = new LogLinkedList();
+    this.baseChunkLogger.clearLogs()
+  }
+
+  public getLogArray() {
+    return this.baseChunkLogger.getLogArray()
   }
 
   public logsToBlob(extraMetadata: Record<string, string>) {
-    this.logs.append(JSON.stringify({ extraMetadata }));
-    const blob = new Blob(this.logs.toArray(), { type: "application/json" });
-    return blob;
+    return this.baseChunkLogger.logsToBlob(extraMetadata)
   }
 }
